Day 1 Part 1 - Sonar Sweeper for small changes
I had some struggles with reduce due to setting a new elevation at the
end of the block. It looks like that new elevation was actually getting
returned and added to the increases. Instead I went ahead and created a
new list that was offset by just one and went through each index,
assembling the number of changes that way instead. Much cleaner I guess?