Day 2 - Navigating the seas with different instructions (#13)
* Day 2 Part 1 - add instructions and inputs

Add answer for part 1

Use a case statement to read each instruction and navigate, then add a
final method to compute the product.

* Add part 2 solution and introduce Command class

Add aim as an attribute and include separate navigation instructions for
part 2. Also introduce a Command class to do the parsing of the "raw
command" separately.

More DRY-ing that could be done here, but this will do for now!