Day 3 Part 1 - A lot more than expected
Came up with the idea for this only in my head and, not surprisingly, it
was a little more challenging than expected to translate into code.

I only wanted to have to loop through the whole file once so came up
with the idea of storing the counts for each position on the way
through. I can't quite tell if that introduced too much complexity, but
at least I got the right answer.