Day 7 Final cleanup - yield to being content
Get the two integers surrounding the floating point mean position and
map to the smaller of the two to calculate the most efficient fuel.
After attempting to use a yield keyword in the same way that Python
does, it became apparent that Ruby uses this differently. While it would
be great to dive into it more, that will be reserved for another day
while these results are yielded to.