2020 - Ruby - Day 6 - Answer the questions! (#19)
* Add instructions, input, and test input

* Add skeleton of specs and main file

* wip

* 2020 - Day 6 part 1 answer

Fun with uniq and join and each_char, oh my!

* Part 2 solution - so much reduce

The logic for this one slowly came together in my head away from the
keyboard. Back at the keyboard, it took me a little while to get this
in the most idiomatic Ruby. I THINK this is the first time I've used
reduce twice in one function :shrug: