2021 - Ruby - Day 7 - Move those crabs! (#20)
* Day 7 Part 1 solution - ~mean~ median position for the win

* Day 7 Part 2 raw solution

Still working out the logic for getting the optimum final position for
the second part, but ended up iterating around the mean and grabbing the
smallest quantity of fuel. Going to work on cleaning it up next.

* Day 7 Final cleanup - yield to being content

Get the two integers surrounding the floating point mean position and
map to the smaller of the two to calculate the most efficient fuel.
After attempting to use a yield keyword in the same way that Python
does, it became apparent that Ruby uses this differently. While it would
be great to dive into it more, that will be reserved for another day
while these results are yielded to.