Ridiculous part 1 solution
I had to use a debugger to get the "paths_hash" from the main example,
then put it into a spec and read the answer there. For some reason the
main program will not spit out the correct answer :shrug:

It may be some bug in the ordering of the hash or something. Or more
likely I am not covering some edge case correctly that the test does
cover for some reason.