Part two non-solution - gotta throw in the towel for now
My part one implementation just squeeked by, but I think it left part
two open to problems. I tried switching "visited" to a hash, but it is
still getting reset after each loop through connections, even if we are
still in the middle of a recursive loop. It should only get reset when
we are totally done. I just can't get it for now and need to set it
aside.