Rust - 2019 - day two - part one - a 1202 program alarm (#30)
* Rust - 2019 - Day 2 - WIP just trying to read the input!

* wow such a long way to go

* Run rustfmt on src/main.rs for day 2 Rust 2019

* Run the failing test, though this needs more work

* WIP working on better Intcode logic

Holy cow this is a challenge with Rust. Gonna keep picking at it.

* Raw but seemingly working day 2 part 1 solution-ish

* Rust day 2 2019 - found the code I guess?

Raw raw solution to get part one for day two. Eager to figure out how to
refactor some of this, but glad for how this solution came together.

* Start cleanup of part 1

* Add part two instructions

* Begin pulling logic out into structs, functions

* Cleaner solution pre-formatting

* Run rustfmt on cleaner solution

* Last part one refactor

I like where this ended up with using the mutable Input value alongside
the instructions that can handle some of the logic of processing each
one.

* Revert "Add part two instructions"

This reverts commit af0a27d875b6b6610bc7a9369acede89491eebd9 so that the
part two instructions can be added later.

* Move part one answer into separate function