Add a template for solving problems in Ruby (#37)
* Add a template for solving in Ruby

Add a template with specs and input files. Define two placeholder
classes: one top-level class to track progress and a lower level class
that could be instantiated as part of the processing.

Eventually a script could be added to copy everything into the correct
directory. Not sure if that's worth the time today =)

Specs are passing by default. The idea is that the new answer could be
plugged into the specs right away with the failing "solve" method
pointed out by the test failure.

* Add a script that will create the directories for the template