Start Rust template and really had a time of it (#38)
* Start Rust template and really had a time of it

Begin a template for Rust questions. Began creating a separate struct
for each type of file input and quickly got stuck trying to turn lines
of comma-separated integers into Vec<Vec<usize>>. Finally got it after
lots of iteration and debugging.

* Add an executable to start new puzzle

Copy paste the script added for Ruby via #37 for use with Rust until a
centralized script can be pulled together. Obviously more work needs to
be done to clean this up.

* Finish with double line break struct example

Create a struct with a from_file method that can work when there are
double line breaks distinguishing classes. This is similar to day one of
2022.