Ruby 2022 - Day four - cleanup time - raw solution (#41)
* Ruby 2022 day four raw solution

Couldn't quite get reduce to work. Wanted to see if I could get to the
answer more quickly without adding the lower-level class. May end up
adding it to make the solution a little cleaner.

* Ruby 2022 day four part two raw solution

Played around with the comparison operators after getting the wrong
answer and drawing out the scenarios. Hoping to make this a little
easier to read/understand later, but for now this will suffice.