Ruby 2022 day five solution - pretty raw
Decided not to write a parser to get the initial stack arrangements and
just hard code them in. It didn't totally feel worth the effort,
especially since I didn't have a ton of time to work on this today and
want to keep moving. Hoping to come back and revisit it, but the input
file also doesn't necessarily need to be read exactly as-is :shrug: