Kotlin 2020 - Day one - Oh, I remember this (#45)
* Initial commit for setting up Kotlin

Includes a .gitignore file sourced from here:

https://github.com/JetBrains/kotlin/blob/master/.gitignore

* Add day1 2020 function

* Kotlin - Day one part one answer with a test

First time running a program or test in Kotlin. Use the test input file
in the unit test and the main file in the main function.

Also update to the latest version of the JVM in the Gradle build
configuration file.

* Kotlin - 2020 day one part two solution

Refactor the file reading out to another method and use it in part two,
which builds off part one and "just" adds another check.