Rust 2023 day two how about some more parsing (#48)
* Rust 2023 Day two part two - max cubes

Yeesh, finally came to this with some messy-looking code that could
probably be refactored into some more succinct match statements. Hoping
to do that soon.

* A little refactoring and cleanup

There may be more that could be done here, but at least this gets rid of
the repetitive code checking each color and instead stores the max
values in a local HashMap.